Delay calls to mutating methods in QScore
Some calls to methods in QScore are initiaied by an event on a
QNote/QMeasure/Qtaff, etc. The resulting operations may end up
deleting the originating graphical object, before it has finished
handling the original event. This causes an error with the mouse
grabbing behaviour.

Fixed this error by decorating certain methods with a delaying method.
This schedules the call to be made after the event handling is finished
by using PyQt's QTimer.singleShot method.

This change also fixes the problem with clicks being ignored after
a QEditMeasureDialog is shown.