Do not use QNotes to display notes. Draw them directly in the QMeasure.
Made a start on translating everything from QNotes to QMeasure. Notes
are now drawn directly in the paint method of QMeasure, which is now
a QGraphicsItem instead of a QGraphicsItemGroup. The QMeasure also
paints the count and the repeat number directly too, so QCount is also not
needed any more.

This change constitutes mostly just moving the machinery of displaying
notes from QNote to QMeasure. It is now much faster than previously,
especially when a big change is made, which used to cause occasional
freezes. Most of the functionality still remains to be ported from QNote
and QCount to QMeasure though. So far, the only operation which has
been ported is toggling a note.