Reworked time counting to be much more flexible.
Changed the back end so that each Beat is an object, which has a Count
indicating the subdivisions of that Beat. A Beat does not have to use
all of a Count if it is incomplete, e.g. the last half-beat of a bar of
7/8.

Each Measure has a MeasureCount, which is a list of Beats.

Fixes #26
Fixes #89

Will make #18, #35 and #88 possible by providing the necessary data
structures.

This was a pretty major change to the back-end data structures, the first
in a while. It entails a slight change to the file format, and the first
time backwards-compatibility has mattered. Scores with old-style time
counting will still work. DrumBurp knows how to create the necessary
Beat/Count for old Scores.

Also implemented an indentation scheme for output files to allow for
easier human readability.