Basis of state machine approach for GUI.
A FSM allows us to reason much more clearly about what we are currently
doing in the GUI, and what actions we can/should take based on input
from the user. This commit is the basis of such an approach. Most of the
previous functionality is converted to using the FSM. Only a couple of
QMeasure operations are left to be changed.